refactor render loop
generalizing the render loop to prototype a moving boid manually
controlled boid
